# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manga Collection API",
        "description": "Manga Collection API Docs",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "local server"
        },
        {
            "url": "https://pink-buffalo-749567.hostingersite.com/api",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/publishers": {
            "get": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Index",
                "operationId": "cf20e994541cc76783eb004350c51d4d",
                "parameters": [
                    {
                        "name": "ordering",
                        "in": "query",
                        "description": "Parameter to order results",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Parameter that set the page",
                        "required": false
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Parameter that set the size of the result collection",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of publishers"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Store",
                "operationId": "a3f49897d9714e054d19ad1b74b963b7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Panini Mangas"
                                    },
                                    "publisher_link": {
                                        "type": "string",
                                        "example": "https://paninimangas.com.br"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Publisher created successfully"
                    },
                    "422": {
                        "description": "Field Error"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "499": {
                        "description": "Not allowed"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/publishers/{publisher_id}": {
            "get": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Show",
                "operationId": "54c08e62846070a945eacee1c3805f3d",
                "parameters": [
                    {
                        "name": "publisher_id",
                        "in": "path",
                        "description": "Parameter that filter the entity",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the requested publisher"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Destroy",
                "operationId": "e40804732ba28ba811bdbd088c065fa7",
                "parameters": [
                    {
                        "name": "publisher_id",
                        "in": "path",
                        "description": "Parameter that filter the entity",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Publisher deleted successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Update",
                "operationId": "a99b2be5143bf8eb3c08c17c9af9eef4",
                "parameters": [
                    {
                        "name": "publisher_id",
                        "in": "path",
                        "description": "Parameter that filter the entity",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "publisher_link": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Publisher updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "499": {
                        "description": "Not allowed"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "Not allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}